Use findup to traverse tree for .git, package.json
Fixes #35, #28, #3. Closes #36

This uses the [findup] package to traverse up the directory tree of
the current working directory, to find `package.json`, or `.git`.
It does so to allow for edge-case situations, such as described in
and #35 (executing hooks from a different working dir, for example
with `git --work-tree`).

As part of this commit, the JSHint maxcomplexity/maxstatements
numbers have been bumped by 1.

[findup]: https://www.npmjs.com/package/findup