chore: Add opt-cli and opt-into the githooks
We want to make contributing to this repository as friendly as possible
So we'll be making the githooks opt-in so people can contribute without
having the barrier of the hooks. But if they start contributing
regularly they can turn them on to avoid the churn of PRs with styling
things.